Make TestModal file to test modal code
